snake, #ifdef usart_printf. Og noen fixes, se changelog

+++ node_1/menu.h
@@ -3,6 +3,7 @@
 #ifndef MENU_H
 #define MENU_H
 
+#include <stdint.h>
 void menu_init();
 void menu_draw(void);
 void menu_main();
@@ -11,7 +12,7 @@ void menu_increment_branch();
 
 void menu_decrement_branch();
 
-int get_menu_branch();
+int8_t get_menu_branch();
 
 void menu_next_page();
 void menu_previous_page();
